Adapt and add rpc_legacy example to book. Add missing light_client example too
