+++ subxt/src/book/usage/mod.rs
@@ -12,6 +12,7 @@
 //! - [Runtime APIs](runtime_apis)
 //! - [Unstable Light Client](light_client)
 //! - [Custom Values](custom_values)
+//! - [RPC calls](rpc)
 //!
 //! Alternately, [go back](super).
 
@@ -23,3 +24,4 @@ pub mod light_client;
 pub mod runtime_apis;
 pub mod storage;
 pub mod transactions;
+pub mod rpc;
